requirements.txt
amqp==2.3.2
astroid==1.6.5
Babel==2.6.0
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-celery-beat==1.1.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
flower==0.9.2
gunicorn==19.8.1
idna==2.6
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.7.4
PyJWT==1.6.4
pylint==1.9.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tornado==5.0.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wrapt==1.10.11
